refactor(`test`): separate tests/fixtures and consolidate via utils
