+++ test/fixtures/required/fixture-conditional-required-multiple.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
-import meow from '../../source/index.js';
+import meow from '../../../source/index.js';
 
 const cli = meow({
 	importMeta: import.meta,
@@ -12,7 +12,7 @@ const cli = meow({
 		test: {
 			type: 'number',
 			shortFlag: 't',
-			isRequired: true,
+			isRequired: () => false,
 			isMultiple: true,
 		},
 	},
