+++ test/test.js
@@ -1,26 +1,14 @@
 import path from 'node:path';
 import process from 'node:process';
-import {fileURLToPath} from 'node:url';
 import test from 'ava';
 import indentString from 'indent-string';
-import {stripIndent} from 'common-tags';
 import {execa} from 'execa';
 import {readPackage} from 'read-pkg';
 import meow from '../source/index.js';
+import {spawnFixture, __dirname} from './_utils.js';
 
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
-const fixturePath = path.join(__dirname, 'fixtures', 'fixture.js');
 const importMeta = import.meta;
 const NODE_MAJOR_VERSION = process.versions.node.split('.')[0];
-
-test('invalid package url', t => {
-	const error = t.throws(() => {
-		meow({
-			importMeta: '/path/to/package',
-		});
-	});
-	t.is(error.message, 'The `importMeta` option is required. Its value must be `import.meta`.');
-});
 
 test('return object', t => {
 	const cli = meow({
@@ -46,64 +34,39 @@ test('return object', t => {
 	t.is(cli.help, indentString('\nCLI app helper\n\nUsage\n  foo <input>\n', 2));
 });
 
-test('support help shortcut', t => {
-	const cli = meow(`
-		unicorn
-		cat
-	`, {
-		importMeta,
-	});
-	t.is(cli.help, indentString('\nCLI app helper\n\nunicorn\ncat\n', 2));
-});
-
 test('spawn cli and show version', async t => {
 	const pkg = await readPackage();
-	const {stdout} = await execa(fixturePath, ['--version']);
+	const {stdout} = await spawnFixture(['--version']);
 	t.is(stdout, pkg.version);
 });
 
 test('spawn cli and disabled autoVersion and autoHelp', async t => {
-	const {stdout} = await execa(fixturePath, ['--version', '--help']);
+	const {stdout} = await spawnFixture(['--version', '--help']);
 	t.is(stdout, 'version\nhelp\nmeow\ncamelCaseOption');
 });
 
 test('spawn cli and disabled autoVersion', async t => {
-	const {stdout} = await execa(fixturePath, ['--version', '--no-auto-version']);
+	const {stdout} = await spawnFixture(['--version', '--no-auto-version']);
 	t.is(stdout, 'version\nautoVersion\nmeow\ncamelCaseOption');
 });
 
 test('spawn cli and not show version', async t => {
-	const {stdout} = await execa(fixturePath, ['--version=beta']);
+	const {stdout} = await spawnFixture(['--version=beta']);
 	t.is(stdout, 'version\nmeow\ncamelCaseOption');
 });
 
-test('spawn cli and show help screen', async t => {
-	const {stdout} = await execa(fixturePath, ['--help']);
-	t.is(stdout, indentString('\nCustom description\n\nUsage\n  foo <input>\n\n', 2));
-});
-
-test('spawn cli and disabled autoHelp', async t => {
-	const {stdout} = await execa(fixturePath, ['--help', '--no-auto-help']);
-	t.is(stdout, 'help\nautoHelp\nmeow\ncamelCaseOption');
-});
-
-test('spawn cli and not show help', async t => {
-	const {stdout} = await execa(fixturePath, ['--help=all']);
-	t.is(stdout, 'help\nmeow\ncamelCaseOption');
-});
-
 test('spawn cli and test input', async t => {
-	const {stdout} = await execa(fixturePath, ['-u', 'cat']);
+	const {stdout} = await spawnFixture(['-u', 'cat']);
 	t.is(stdout, 'unicorn\nmeow\ncamelCaseOption');
 });
 
 test('spawn cli and test input flag', async t => {
-	const {stdout} = await execa(fixturePath, ['--camel-case-option', 'bar']);
+	const {stdout} = await spawnFixture(['--camel-case-option', 'bar']);
 	t.is(stdout, 'bar');
 });
 
 test('spawn cli and test process title', async t => {
-	const {stdout} = await execa(path.join(__dirname, 'fixtures/with-package-json/fixture.js'));
+	const {stdout} = await spawnFixture('with-package-json/fixture.js');
 	t.is(stdout, 'foo');
 });
 
@@ -126,16 +89,6 @@ test('single character flag casing should be preserved', t => {
 		importMeta,
 		argv: ['-F'],
 	}).flags, {F: true});
-});
-
-test('flag declared in kebab-case is an error', t => {
-	const error = t.throws(() => {
-		meow({
-			importMeta,
-			flags: {'kebab-case': 'boolean', test: 'boolean', 'another-one': 'boolean'},
-		});
-	});
-	t.is(error.message, 'Flag keys may not contain \'-\'. Invalid flags: `kebab-case`, `another-one`');
 });
 
 test('type inference', t => {
@@ -349,578 +302,6 @@ test('supports `number` flag type - no flag but default', t => {
 	}).flags.foo;
 
 	t.is(cli, 2);
-});
-
-test('supports `number` flag type - throws on incorrect default value', t => {
-	t.throws(() => {
-		meow({
-			importMeta,
-			argv: [],
-			flags: {
-				foo: {
-					type: 'number',
-					default: 'x',
-				},
-			},
-		});
-	});
-});
-
-test('isMultiple - unset flag returns empty array', t => {
-	t.deepEqual(meow({
-		importMeta,
-		argv: [],
-		flags: {
-			foo: {
-				type: 'string',
-				isMultiple: true,
-			},
-		},
-	}).flags, {
-		foo: [],
-	});
-});
-
-test('isMultiple - flag set once returns array', t => {
-	t.deepEqual(meow({
-		importMeta,
-		argv: ['--foo=bar'],
-		flags: {
-			foo: {
-				type: 'string',
-				isMultiple: true,
-			},
-		},
-	}).flags, {
-		foo: ['bar'],
-	});
-});
-
-test('isMultiple - flag set multiple times', t => {
-	t.deepEqual(meow({
-		importMeta,
-		argv: ['--foo=bar', '--foo=baz'],
-		flags: {
-			foo: {
-				type: 'string',
-				isMultiple: true,
-			},
-		},
-	}).flags, {
-		foo: ['bar', 'baz'],
-	});
-});
-
-test('isMultiple - flag with space separated values', t => {
-	const {input, flags} = meow({
-		importMeta,
-		argv: ['--foo', 'bar', 'baz'],
-		flags: {
-			foo: {
-				type: 'string',
-				isMultiple: true,
-			},
-		},
-	});
-
-	t.deepEqual(input, ['baz']);
-	t.deepEqual(flags.foo, ['bar']);
-});
-
-test('isMultiple - flag with comma separated values', t => {
-	t.deepEqual(meow({
-		importMeta,
-		argv: ['--foo', 'bar,baz'],
-		flags: {
-			foo: {
-				type: 'string',
-				isMultiple: true,
-			},
-		},
-	}).flags, {
-		foo: ['bar,baz'],
-	});
-});
-
-test('single flag set more than once => throws', t => {
-	t.throws(() => {
-		meow({
-			importMeta,
-			argv: ['--foo=bar', '--foo=baz'],
-			flags: {
-				foo: {
-					type: 'string',
-				},
-			},
-		});
-	}, {message: 'The flag --foo can only be set once.'});
-});
-
-test('isMultiple - default to type string', t => {
-	t.deepEqual(meow({
-		importMeta,
-		argv: ['--foo=bar'],
-		flags: {
-			foo: {
-				isMultiple: true,
-			},
-		},
-	}).flags, {
-		foo: ['bar'],
-	});
-});
-
-test('isMultiple - boolean flag', t => {
-	t.deepEqual(meow({
-		importMeta,
-		argv: ['--foo', '--foo=false'],
-		flags: {
-			foo: {
-				type: 'boolean',
-				isMultiple: true,
-			},
-		},
-	}).flags, {
-		foo: [true, false],
-	});
-});
-
-test('isMultiple - boolean flag is false by default', t => {
-	t.deepEqual(meow({
-		importMeta,
-		argv: [],
-		flags: {
-			foo: {
-				type: 'boolean',
-				isMultiple: true,
-			},
-		},
-	}).flags, {
-		foo: [false],
-	});
-});
-
-test('isMultiple - number flag', t => {
-	t.deepEqual(meow({
-		importMeta,
-		argv: ['--foo=1.3', '--foo=-1'],
-		flags: {
-			foo: {
-				type: 'number',
-				isMultiple: true,
-			},
-		},
-	}).flags, {
-		foo: [1.3, -1],
-	});
-});
-
-test('isMultiple - flag default values', t => {
-	t.deepEqual(meow({
-		importMeta,
-		argv: [],
-		flags: {
-			string: {
-				type: 'string',
-				isMultiple: true,
-				default: ['foo'],
-			},
-			boolean: {
-				type: 'boolean',
-				isMultiple: true,
-				default: [true],
-			},
-			number: {
-				type: 'number',
-				isMultiple: true,
-				default: [0.5],
-			},
-		},
-	}).flags, {
-		string: ['foo'],
-		boolean: [true],
-		number: [0.5],
-	});
-});
-
-test('isMultiple - multiple flag default values', t => {
-	t.deepEqual(meow({
-		importMeta,
-		argv: [],
-		flags: {
-			string: {
-				type: 'string',
-				isMultiple: true,
-				default: ['foo', 'bar'],
-			},
-			boolean: {
-				type: 'boolean',
-				isMultiple: true,
-				default: [true, false],
-			},
-			number: {
-				type: 'number',
-				isMultiple: true,
-				default: [0.5, 1],
-			},
-		},
-	}).flags, {
-		string: ['foo', 'bar'],
-		boolean: [true, false],
-		number: [0.5, 1],
-	});
-});
-
-// Happened in production 2020-05-10: https://github.com/sindresorhus/meow/pull/143#issuecomment-626287226
-test('isMultiple - handles multi-word flag name', t => {
-	t.deepEqual(meow({
-		importMeta,
-		argv: ['--foo-bar=baz'],
-		flags: {
-			fooBar: {
-				type: 'string',
-				isMultiple: true,
-			},
-		},
-	}).flags, {
-		fooBar: ['baz'],
-	});
-});
-
-test('suggests renaming alias to shortFlag', t => {
-	t.throws(() => {
-		meow({
-			importMeta,
-			flags: {
-				foo: {
-					type: 'string',
-					alias: 'f',
-				},
-				bar: {
-					type: 'string',
-					alias: 'b',
-				},
-				baz: {
-					type: 'string',
-					shortFlag: 'z',
-				},
-			},
-		});
-	}, {message: 'The option `alias` has been renamed to `shortFlag`. The following flags need to be updated: `--foo`, `--bar`'});
-});
-
-test('aliases - accepts one', t => {
-	t.deepEqual(meow({
-		importMeta,
-		argv: ['--foo=baz'],
-		flags: {
-			fooBar: {
-				type: 'string',
-				aliases: ['foo'],
-			},
-		},
-	}).flags, {
-		fooBar: 'baz',
-	});
-});
-
-test('aliases - accepts multiple', t => {
-	t.deepEqual(meow({
-		importMeta,
-		argv: ['--foo=baz1', '--bar=baz2'],
-		flags: {
-			fooBar: {
-				type: 'string',
-				aliases: ['foo', 'bar'],
-				isMultiple: true,
-			},
-		},
-	}).flags, {
-		fooBar: ['baz1', 'baz2'],
-	});
-});
-
-test('aliases - can be a short flag', t => {
-	t.deepEqual(meow({
-		importMeta,
-		argv: ['--f=baz'],
-		flags: {
-			fooBar: {
-				type: 'string',
-				aliases: ['f'],
-			},
-		},
-	}).flags, {
-		fooBar: 'baz',
-	});
-});
-
-test('aliases - works with short flag', t => {
-	t.deepEqual(meow({
-		importMeta,
-		argv: ['--foo=baz1', '--bar=baz2', '-f=baz3'],
-		flags: {
-			fooBar: {
-				type: 'string',
-				shortFlag: 'f',
-				aliases: ['foo', 'bar'],
-				isMultiple: true,
-			},
-		},
-	}).flags, {
-		fooBar: ['baz1', 'baz2', 'baz3'],
-	});
-});
-
-test('aliases - unnormalized flags', t => {
-	t.deepEqual(meow({
-		importMeta,
-		argv: ['--foo=baz'],
-		flags: {
-			fooBar: {
-				type: 'string',
-				aliases: ['foo'],
-				shortFlag: 'f',
-			},
-		},
-	}).unnormalizedFlags, {
-		fooBar: 'baz',
-		foo: 'baz',
-		f: 'baz',
-	});
-});
-
-test('choices - success case', t => {
-	const cli = meow({
-		importMeta,
-		argv: ['--animal', 'cat', '--number=2.2'],
-		flags: {
-			animal: {
-				choices: ['dog', 'cat', 'unicorn'],
-			},
-			number: {
-				type: 'number',
-				choices: [1.1, 2.2, 3.3],
-			},
-		},
-	});
-
-	t.is(cli.flags.animal, 'cat');
-	t.is(cli.flags.number, 2.2);
-});
-
-test('choices - throws if input does not match choices', t => {
-	t.throws(() => {
-		meow({
-			importMeta,
-			argv: ['--animal', 'rainbow', '--number', 5],
-			flags: {
-				animal: {
-					choices: ['dog', 'cat', 'unicorn'],
-				},
-				number: {
-					choices: [1, 2, 3],
-				},
-			},
-		});
-	}, {
-		message: stripIndent`
-			Unknown value for flag \`--animal\`: \`rainbow\`. Value must be one of: [\`dog\`, \`cat\`, \`unicorn\`]
-			Unknown value for flag \`--number\`: \`5\`. Value must be one of: [\`1\`, \`2\`, \`3\`]
-		`,
-	});
-});
-
-test('choices - throws if choices is not array', t => {
-	t.throws(() => {
-		meow({
-			importMeta,
-			argv: ['--animal', 'cat'],
-			flags: {
-				animal: {
-					choices: 'cat',
-				},
-			},
-		});
-	}, {message: 'The option `choices` must be an array. Invalid flags: `--animal`'});
-});
-
-test('choices - does not throw error when isRequired is false', t => {
-	t.notThrows(() => {
-		meow({
-			importMeta,
-			argv: [],
-			flags: {
-				animal: {
-					isRequired: false,
-					choices: ['dog', 'cat', 'unicorn'],
-				},
-			},
-		});
-	});
-});
-
-test('choices - throw error when isRequired is true', t => {
-	t.throws(() => {
-		meow({
-			importMeta,
-			argv: [],
-			flags: {
-				animal: {
-					isRequired: true,
-					choices: ['dog', 'cat', 'unicorn'],
-				},
-			},
-		});
-	}, {message: 'Flag `--animal` has no value. Value must be one of: [`dog`, `cat`, `unicorn`]'});
-});
-
-test('choices - success with isMultiple', t => {
-	const cli = meow({
-		importMeta,
-		argv: ['--animal=dog', '--animal=unicorn'],
-		flags: {
-			animal: {
-				type: 'string',
-				isMultiple: true,
-				choices: ['dog', 'cat', 'unicorn'],
-			},
-		},
-	});
-
-	t.deepEqual(cli.flags.animal, ['dog', 'unicorn']);
-});
-
-test('choices - throws with isMultiple, one unknown value', t => {
-	t.throws(() => {
-		meow({
-			importMeta,
-			argv: ['--animal=dog', '--animal=rabbit'],
-			flags: {
-				animal: {
-					type: 'string',
-					isMultiple: true,
-					choices: ['dog', 'cat', 'unicorn'],
-				},
-			},
-		});
-	}, {message: 'Unknown value for flag `--animal`: `rabbit`. Value must be one of: [`dog`, `cat`, `unicorn`]'});
-});
-
-test('choices - throws with isMultiple, multiple unknown value', t => {
-	t.throws(() => {
-		meow({
-			importMeta,
-			argv: ['--animal=dog', '--animal=rabbit'],
-			flags: {
-				animal: {
-					type: 'string',
-					isMultiple: true,
-					choices: ['cat', 'unicorn'],
-				},
-			},
-		});
-	}, {message: 'Unknown values for flag `--animal`: `dog`, `rabbit`. Value must be one of: [`cat`, `unicorn`]'});
-});
-
-test('choices - throws with multiple flags', t => {
-	t.throws(() => {
-		meow({
-			importMeta,
-			argv: ['--animal=dog', '--plant=succulent'],
-			flags: {
-				animal: {
-					type: 'string',
-					choices: ['cat', 'unicorn'],
-				},
-				plant: {
-					type: 'string',
-					choices: ['tree', 'flower'],
-				},
-			},
-		});
-	}, {message: stripIndent`
-		Unknown value for flag \`--animal\`: \`dog\`. Value must be one of: [\`cat\`, \`unicorn\`]
-		Unknown value for flag \`--plant\`: \`succulent\`. Value must be one of: [\`tree\`, \`flower\`]
-	`});
-});
-
-test('choices - choices must be of the same type', t => {
-	t.throws(() => {
-		meow({
-			importMeta,
-			flags: {
-				number: {
-					type: 'number',
-					choices: [1, '2'],
-				},
-				boolean: {
-					type: 'boolean',
-					choices: [true, 'false'],
-				},
-			},
-		});
-	}, {message: 'Each value of the option `choices` must be of the same type as its flag. Invalid flags: (`--number`, type: \'number\'), (`--boolean`, type: \'boolean\')'});
-});
-
-test('choices - default must only include valid choices', t => {
-	t.throws(() => {
-		meow({
-			importMeta,
-			flags: {
-				number: {
-					type: 'number',
-					choices: [1, 2, 3],
-					default: 1,
-				},
-			},
-		});
-	}, {message: 'Each value of the option `default` must exist within the option `choices`. Invalid flags: `--number`'});
-});
-
-test('options - multiple validation errors', t => {
-	t.throws(() => {
-		meow({
-			importMeta,
-			flags: {
-				animal: {
-					type: 'string',
-					choices: 'cat',
-				},
-				plant: {
-					type: 'string',
-					alias: 'p',
-				},
-				'some-thing': {
-					type: 'string',
-				},
-			},
-		});
-	}, {message: stripIndent`
-		Flag keys may not contain '-'. Invalid flags: \`some-thing\`
-		The option \`alias\` has been renamed to \`shortFlag\`. The following flags need to be updated: \`--plant\`
-		The option \`choices\` must be an array. Invalid flags: \`--animal\`
-	`});
-});
-
-test('single line help messages are not indented', t => {
-	const cli = meow({
-		importMeta,
-		description: false,
-		help: 'single line',
-	});
-
-	t.is(cli.help, '\nsingle line\n');
-});
-
-test('descriptions with no help are not indented', t => {
-	const cli = meow({
-		importMeta,
-		help: false,
-		description: 'single line',
-	});
-
-	t.is(cli.help, '\nsingle line\n');
 });
 
 if (NODE_MAJOR_VERSION >= 14) {
